+++ tensor/internal/validator/reducers.go
@@ -2,8 +2,8 @@ package validator
 
 import "fmt"
 
-func ValidateReducedDimAgainstDims(dim int32, dims []int32) (err error) {
-	if !(0 <= dim && dim < int32(len(dims))) {
+func ValidateReducedDimAgainstDims(dim int, dims []int) (err error) {
+	if !(0 <= dim && dim < len(dims)) {
 		err = fmt.Errorf("expected dimension to be in range [0,%d): got (%d)", len(dims), dim)
 		return
 	}
